Add Features Chat Bot, Add Shipper table and Shipping Company management

+++ src/main/java/com/dev/shoeshop/repository/CartDetailRepository.java
@@ -12,7 +12,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface CartDetailRepository extends JpaRepository<CartDetail, Long> {
+public interface CartDetailRepository extends JpaRepository<CartDetail, Integer> {
     
     /**
      * Find all cart details by cart
